requirements-development.txt
-r requirements.txt

pytest>=3.3.0               # Testing
pytest-cov>=2.5.0           # Coverage Integration
pytest-django>=3.1.0        # Django Integration

flake8>=3.5.0               # Linting
factory_boy>=2.9.0          # Generating fake data
graphene>=2.0.0             # For testing graphql utilities
